The TCalcTransport is used by a single caller at a time, with a
mutex used to protect access:
https://github.com/uber-go/tally/blob/5431c04afc0bad76ce91ca3d549abb469abd2a10/m3/reporter.go#L400

The thrift generated code will only call `Write` serially, so there's no
need to use atomics here.

Removal of atomics gives a small performance improvement (10-15%):
```
> benchcmp -best old new
benchmark                 old ns/op     new ns/op     delta
BenchmarkCalulateSize     538           453           -15.80%
```